asyncapi-service/pyproject.toml
[tool.poetry]
name = "fastapiservice"
version = "0.1.0"
description = ""
authors = ["Ruslan Filichkin <ruslan.filichkin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
elasticsearch = {extras = ["async"], version = "7.9.1"}
flake8 = "5.0.4"
isort = "5.10.1"
autopep8 = "1.7.0"
pytest = "7.1.2"
pre-commit = "2.20.0"
python-dotenv = "0.20.0"
pydantic = "1.9.1"
fastapi = "0.79.0"
orjson = "3.7.11"
uvicorn = "0.18.2"
uvloop = "0.16.0"
aioredis = "1.3.1"
gunicorn = "20.1.0"
requests = "2.28.1"
backoff = "2.1.2"
pytest-asyncio = "0.12.0"
redis = "4.3.4"
PyJWT = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
